fix: "force choose" is invalid when one value is chosen (#65)

+++ src/util/byte_writer.rs
@@ -37,17 +37,15 @@ impl<const EMPTY: bool> ByteWriter<EMPTY> {
     pub fn write_value<T: AsBytes>(&mut self, value: T) {
         let data = value.as_bytes();
         if !EMPTY {
-            self.buffer[self.index as usize..self.index as usize + data.len()]
-                .copy_from_slice(value.as_bytes());
+            self.buffer[self.index..self.index + data.len()].copy_from_slice(value.as_bytes());
         }
         self.index += data.len();
     }
 
     pub fn write_value_bytes<T: AsBytes>(&mut self, value: T, bytes: usize) {
         if !EMPTY {
-            self.buffer[self.index as usize..self.index as usize + bytes]
-                .copy_from_slice(&value.as_bytes()[..bytes]);
+            self.buffer[self.index..self.index + bytes].copy_from_slice(&value.as_bytes()[..bytes]);
         }
-        self.index += bytes as usize;
+        self.index += bytes;
     }
 }
